refactor code except for edit button
